Fix bug when function definition is at the beginning of the file (#5)
If a function definition is at the beginning of the file, when we try to get the symbol table of the function we instead get the global table. This means the types of the arguments cannot be inferred. This PR fixes the problem by returning the innermost scope of a given position in the source code instead of the outermost scope.